+++ hikka/heroku.py
@@ -120,6 +120,7 @@ def get_repo():
         repo.create_head("master", origin.refs.master)
         repo.heads.master.set_tracking_branch(origin.refs.master)
         repo.heads.master.checkout(True)
+
     return repo