+++ hikka/modules/loader.py
@@ -63,43 +63,49 @@ class LoaderMod(loader.Module):
 
     strings = {
         "name": "Loader",
-        "repo_config_doc": "Fully qualified URL to a module repo",
+        "repo_config_doc": "URL to a module repo",
         "avail_header": (
-            "<emoji document_id='6321352876505434037'>🎢</emoji><b> Official modules"
+            "<emoji document_id='6321352876505434037'>🎢</emoji><b> Modules"
             " from repo</b>"
         ),
         "select_preset": "<b>⚠️ Please select a preset</b>",
         "no_preset": (
-            "<emoji document_id='5415905755406539934'>🚫</emoji><b> Preset not found</b>"
+            "<emoji document_id='5375201396859607943'>🚫</emoji><b> Preset not found</b>"
         ),
         "preset_loaded": (
             "<emoji document_id='6323332130579416910'>✅</emoji><b> Preset loaded</b>"
         ),
         "no_module": (
-            "<emoji document_id='5415905755406539934'>🚫</emoji><b> Module not available"
+            "<emoji document_id='5375201396859607943'>🚫</emoji><b> Module not available"
             " in repo.</b>"
         ),
         "no_file": (
-            "<emoji document_id='5415905755406539934'>🚫</emoji><b> File not found</b>"
+            "<emoji document_id='5375201396859607943'>🚫</emoji><b> File not found</b>"
         ),
         "provide_module": "<b>⚠️ Provide a module to load</b>",
         "bad_unicode": (
-            "<emoji document_id='5415905755406539934'>🚫</emoji><b> Invalid Unicode"
+            "<emoji document_id='5375201396859607943'>🚫</emoji><b> Invalid Unicode"
             " formatting in module</b>"
         ),
         "load_failed": (
-            "<emoji document_id='5415905755406539934'>🚫</emoji><b> Loading failed. See"
+            "<emoji document_id='5375201396859607943'>🚫</emoji><b> Loading failed. See"
             " logs for details</b>"
         ),
-        "loaded": "<b>🔭 Module </b><code>{}</code>{}<b> loaded {}</b>{}{}{}{}{}{}",
+        "loaded": (
+            "<emoji document_id='5188377234380954537'>🌘</emoji><b> Module"
+            " </b><code>{}</code>{}<b> loaded {}</b>{}{}{}{}{}{}"
+        ),
         "no_class": "<b>What class needs to be unloaded?</b>",
-        "unloaded": "<b>🧹 Module {} unloaded.</b>",
+        "unloaded": (
+            "<emoji document_id='5469654973308476699'>💣</emoji><b> Module {}"
+            " unloaded.</b>"
+        ),
         "not_unloaded": (
-            "<emoji document_id='5415905755406539934'>🚫</emoji><b> Module not"
+            "<emoji document_id='5375201396859607943'>🚫</emoji><b> Module not"
             " unloaded.</b>"
         ),
         "requirements_failed": (
-            "<emoji document_id='5415905755406539934'>🚫</emoji><b> Requirements"
+            "<emoji document_id='5375201396859607943'>🚫</emoji><b> Requirements"
             " installation failed</b>"
         ),
         "requirements_failed_termux": (
@@ -112,10 +118,14 @@ class LoaderMod(loader.Module):
             " can't be done on Heroku. Don't report it as bug, this can't be"
             " solved.</b>"
         ),
-        "requirements_installing": "<b>🔄 Installing requirements:\n\n{}</b>",
+        "requirements_installing": (
+            "<emoji document_id='5445284980978621387'>🚀</emoji><b> Installing"
+            " requirements:\n\n{}</b>"
+        ),
         "requirements_restart": (
-            "<b>🔄 Requirements installed, but a restart is required for"
-            " </b><code>{}</code><b> to apply</b>"
+            "<emoji document_id='5445284980978621387'>🚀</emoji><b> Requirements"
+            " installed, but a restart is required for </b><code>{}</code><b> to"
+            " apply</b>"
         ),
         "all_modules_deleted": (
             "<emoji document_id='6323332130579416910'>✅</emoji><b> All modules"
@@ -140,11 +150,11 @@ class LoaderMod(loader.Module):
             " FFMPEG, which is not installed</b>"
         ),
         "developer": (
-            "\n\n<emoji document_id='5444965061749644170'>👨‍💻</emoji> <b>Developer:"
+            "\n\n<emoji document_id='5431376038628171216'>👨‍💻</emoji> <b>Developer:"
             " </b>{}"
         ),
         "depends_from": (
-            "\n\n<emoji document_id='5458549695414543088'>📦</emoji> <b>Dependencies:"
+            "\n\n<emoji document_id='5431736674147114227'>📦</emoji> <b>Dependencies:"
             " </b>\n{}"
         ),
         "by": "by",
@@ -182,20 +192,23 @@ class LoaderMod(loader.Module):
         "cancel": "🚫 Cancel",
         "overwrite_module": (
             "<emoji document_id='5454225457916420314'>😖</emoji> <b>This module"
-            " attempted to override the core one (</b><code>{}</code><b>)</b>\n\n<i>💡"
-            " Don't report it as bug. It's a security measure to prevent replacing core"
-            " modules with some junk</i>"
+            " attempted to override the core one (</b><code>{}</code><b>)</b>\n\n<emoji"
+            " document_id='5472146462362048818'>💡</emoji><i> Don't report it as bug."
+            " It's a security measure to prevent replacing core modules with some"
+            " junk</i>"
         ),
         "overwrite_command": (
             "<emoji document_id='5454225457916420314'>😖</emoji> <b>This module"
             " attempted to override the core command"
-            " (</b><code>{}{}</code><b>)</b>\n\n<i>💡 Don't report it as bug. It's a"
-            " security measure to prevent replacing core modules' commands with some"
-            " junk</i>"
+            " (</b><code>{}{}</code><b>)</b>\n\n<emoji"
+            " document_id='5472146462362048818'>💡</emoji><i> Don't report it as bug."
+            " It's a security measure to prevent replacing core modules' commands with"
+            " some junk</i>"
         ),
         "unload_core": (
             "<emoji document_id='5454225457916420314'>😖</emoji> <b>You can't unload"
-            " core module </b><code>{}</code><b></b>\n\n<i>💡 Don't report it as bug."
+            " core module </b><code>{}</code><b></b>\n\n<emoji"
+            " document_id='5472146462362048818'>💡</emoji><i> Don't report it as bug."
             " It's a security measure to prevent replacing core modules with some"
             " junk</i>"
         ),
@@ -204,10 +217,10 @@ class LoaderMod(loader.Module):
             " library</b>"
         ),
         "wait_channel_approve": (
-            "<emoji document_id='5384578263949515741'>💫</emoji> <b>Module"
+            "<emoji document_id='5469741319330996757'>💫</emoji> <b>Module"
             " </b><code>{}</code><b> requests permission to join channel <a"
             ' href="https://t.me/{}">{}</a>.\n\n<b><emoji'
-            ' document_id="6037389971269094179">❓</emoji> Reason: {}</b>\n\n<i>Waiting'
+            ' document_id="5467666648263564704">❓</emoji> Reason: {}</b>\n\n<i>Waiting'
             ' for <a href="https://t.me/{}">approval</a>...</i>'
         ),
     }
@@ -221,36 +234,42 @@ class LoaderMod(loader.Module):
         ),
         "select_preset": "<b>⚠️ Выбери пресет</b>",
         "no_preset": (
-            "<emoji document_id='5415905755406539934'>🚫</emoji><b> Пресет не найден</b>"
+            "<emoji document_id='5375201396859607943'>🚫</emoji><b> Пресет не найден</b>"
         ),
         "preset_loaded": (
             "<emoji document_id='6323332130579416910'>✅</emoji><b> Пресет загружен</b>"
         ),
         "no_module": (
-            "<emoji document_id='5415905755406539934'>🚫</emoji><b> Модуль недоступен в"
+            "<emoji document_id='5375201396859607943'>🚫</emoji><b> Модуль недоступен в"
             " репозитории.</b>"
         ),
         "no_file": (
-            "<emoji document_id='5415905755406539934'>🚫</emoji><b> Файл не найден</b>"
+            "<emoji document_id='5375201396859607943'>🚫</emoji><b> Файл не найден</b>"
         ),
         "provide_module": "<b>⚠️ Укажи модуль для загрузки</b>",
         "bad_unicode": (
-            "<emoji document_id='5415905755406539934'>🚫</emoji><b> Неверная кодировка"
+            "<emoji document_id='5375201396859607943'>🚫</emoji><b> Неверная кодировка"
             " модуля</b>"
         ),
         "load_failed": (
-            "<emoji document_id='5415905755406539934'>🚫</emoji><b> Загрузка не"
+            "<emoji document_id='5375201396859607943'>🚫</emoji><b> Загрузка не"
             " увенчалась успехом. Смотри логи.</b>"
         ),
-        "loaded": "<b>🔭 Модуль </b><code>{}</code>{}<b> загружен {}</b>{}{}{}{}{}{}",
+        "loaded": (
+            "<emoji document_id='5188377234380954537'>🌘</emoji><b> Модуль"
+            " </b><code>{}</code>{}<b> загружен {}</b>{}{}{}{}{}{}"
+        ),
         "no_class": "<b>А что выгружать то?</b>",
-        "unloaded": "<b>🧹 Модуль {} выгружен.</b>",
+        "unloaded": (
+            "<emoji document_id='5469654973308476699'>💣</emoji><b> Модуль {}"
+            " выгружен.</b>"
+        ),
         "not_unloaded": (
-            "<emoji document_id='5415905755406539934'>🚫</emoji><b> Модуль не"
+            "<emoji document_id='5375201396859607943'>🚫</emoji><b> Модуль не"
             " выгружен.</b>"
         ),
         "requirements_failed": (
-            "<emoji document_id='5415905755406539934'>🚫</emoji><b> Ошибка установки"
+            "<emoji document_id='5375201396859607943'>🚫</emoji><b> Ошибка установки"
             " зависимостей</b>"
         ),
         "requirements_failed_termux": (
@@ -264,10 +283,13 @@ class LoaderMod(loader.Module):
             " установить на Heroku. Не сообщайте об этом как об ошибке, это не может"
             " быть исправлено</b>"
         ),
-        "requirements_installing": "<b>🔄 Устанавливаю зависимости:\n\n{}</b>",
+        "requirements_installing": (
+            "<emoji document_id='5445284980978621387'>🚀</emoji><b> Устанавливаю"
+            " зависимости:\n\n{}</b>"
+        ),
         "requirements_restart": (
-            "<b>🔄 Зависимости установлены, но нужна перезагрузка для применения"
-            " </b><code>{}</code>"
+            "<emoji document_id='5445284980978621387'>🚀</emoji><b> Зависимости"
+            " установлены, но нужна перезагрузка для применения </b><code>{}</code>"
         ),
         "all_modules_deleted": (
             "<emoji document_id='6323332130579416910'>✅</emoji><b> Модули удалены</b>"
@@ -285,11 +307,11 @@ class LoaderMod(loader.Module):
             " требуется FFMPEG, который не установлен</b>"
         ),
         "developer": (
-            "\n\n<emoji document_id='5444965061749644170'>👨‍💻</emoji> <b>Разработчик:"
+            "\n\n<emoji document_id='5431376038628171216'>👨‍💻</emoji> <b>Разработчик:"
             " </b>{}"
         ),
         "depends_from": (
-            "\n\n<emoji document_id='5458549695414543088'>📦</emoji> <b>Зависимости:"
+            "\n\n<emoji document_id='5431736674147114227'>📦</emoji> <b>Зависимости:"
             " </b>\n{}"
         ),
         "by": "от",
@@ -344,32 +366,35 @@ class LoaderMod(loader.Module):
         "cancel": "🚫 Отмена",
         "overwrite_module": (
             "<emoji document_id='5454225457916420314'>😖</emoji> <b>Этот модуль"
-            " попытался перезаписать встроенный (</b><code>{}</code><b>)</b>\n\n<i>💡"
-            " Это не ошибка, а мера безопасности, требуемая для предотвращения замены"
-            " встроенных модулей всяким хламом. Не сообщайте о ней в support чате</i>"
+            " попытался перезаписать встроенный (</b><code>{}</code><b>)</b>\n\n<emoji"
+            " document_id='5472146462362048818'>💡</emoji><i> Это не ошибка, а мера"
+            " безопасности, требуемая для предотвращения замены встроенных модулей"
+            " всяким хламом. Не сообщайте о ней в support чате</i>"
         ),
         "overwrite_command": (
             "<emoji document_id='5454225457916420314'>😖</emoji> <b>Этот модуль"
             " попытался перезаписать встроенную команду"
-            " (</b><code>{}</code><b>)</b>\n\n<i>💡 Это не ошибка, а мера безопасности,"
-            " требуемая для предотвращения замены команд встроенных модулей всяким"
-            " хламом. Не сообщайте о ней в support чате</i>"
+            " (</b><code>{}</code><b>)</b>\n\n<emoji"
+            " document_id='5472146462362048818'>💡</emoji><i> Это не ошибка, а мера"
+            " безопасности, требуемая для предотвращения замены команд встроенных"
+            " модулей всяким хламом. Не сообщайте о ней в support чате</i>"
         ),
         "unload_core": (
             "<emoji document_id='5454225457916420314'>😖</emoji> <b>Ты не можешь"
-            " выгрузить встроенный модуль </b><code>{}</code><b></b>\n\n<i>💡 Это не"
-            " ошибка, а мера безопасности, требуемая для предотвращения замены"
-            " встроенных модулей всяким хламом. Не сообщайте о ней в support чате</i>"
+            " выгрузить встроенный модуль </b><code>{}</code><b></b>\n\n<emoji"
+            " document_id='5472146462362048818'>💡</emoji><i> Это не ошибка, а мера"
+            " безопасности, требуемая для предотвращения замены встроенных модулей"
+            " всяким хламом. Не сообщайте о ней в support чате</i>"
         ),
         "cannot_unload_lib": (
             "<emoji document_id='5454225457916420314'>😖</emoji> <b>Ты не можешь"
             " выгрузить библиотеку</b>"
         ),
         "wait_channel_approve": (
-            "<emoji document_id='5384578263949515741'>💫</emoji> <b>Модуль"
+            "<emoji document_id='5469741319330996757'>💫</emoji> <b>Модуль"
             " </b><code>{}</code><b> запрашивает разрешение на вступление в канал <a"
             ' href="https://t.me/{}">{}</a>.\n\n<b><emoji'
-            ' document_id="6037389971269094179">❓</emoji> Причина:'
+            ' document_id="5467666648263564704">❓</emoji> Причина:'
             ' {}</b>\n\n<i>Ожидание <a href="https://t.me/{}">подтверждения</a>...</i>'
         ),
     }
@@ -917,7 +942,7 @@ class LoaderMod(loader.Module):
                     ),
                     origin=f"<string {uid}>" if origin == "<string>" else origin,
                 )
-                instance = self.allmodules.register_module(
+                instance = await self.allmodules.register_module(
                     spec,
                     module_name,
                     origin,
@@ -1012,7 +1037,7 @@ class LoaderMod(loader.Module):
                 return await self.load_module(**kwargs)  # Try again
             except loader.LoadError as e:
                 with contextlib.suppress(Exception):
-                    self.allmodules.unload_module(instance.__class__.__name__)
+                    await self.allmodules.unload_module(instance.__class__.__name__)
 
                 with contextlib.suppress(Exception):
                     self.allmodules.modules.remove(instance)
@@ -1082,7 +1107,7 @@ class LoaderMod(loader.Module):
                 task.cancel()
             except loader.LoadError as e:
                 with contextlib.suppress(Exception):
-                    self.allmodules.unload_module(instance.__class__.__name__)
+                    await self.allmodules.unload_module(instance.__class__.__name__)
 
                 with contextlib.suppress(Exception):
                     self.allmodules.modules.remove(instance)
@@ -1097,7 +1122,7 @@ class LoaderMod(loader.Module):
             except loader.SelfUnload as e:
                 logging.debug(f"Unloading {instance}, because it raised `SelfUnload`")
                 with contextlib.suppress(Exception):
-                    self.allmodules.unload_module(instance.__class__.__name__)
+                    await self.allmodules.unload_module(instance.__class__.__name__)
 
                 with contextlib.suppress(Exception):
                     self.allmodules.modules.remove(instance)
@@ -1327,7 +1352,7 @@ class LoaderMod(loader.Module):
             return
 
         try:
-            worked = self.allmodules.unload_module(args)
+            worked = await self.allmodules.unload_module(args)
         except CoreUnloadError as e:
             await utils.answer(message, self.strings("unload_core").format(e.module))
             return