+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Hikka Changelog
 
+## 🌑 Hikka 1.4.0
+
+- Publish hikka telethon fork and migrate to it in requirements, thereby fixing the deployment error on Heroku
+- Add custom emojis filter to `utils.remove_html`
+- Fix `client.get_perms_cached`
+- Fix translation flaw in `HikkaSecurity`
+- Fix `.uninstall_hikka` being accessible by sudo
+- Fix `utils.find_caller` for :method:`hikka.inline.utils.Utils._find_caller_sec_map`
+- Add more animated emojis to modules
+- Add targeted security for users and chats (`.tsec`)
+- Add support for `tg_level` in `.config Tester`
+
 ## 🌑 Hikka 1.3.3
 
 - Fix typo, which broke `client.get_fulluser`